% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bounds.R
\name{get_default_bounds}
\alias{get_default_bounds}
\title{Default Bounds Initialization for SVR Optimization}
\usage{
get_default_bounds()
}
\value{
A list containing:
\describe{
  \item{lb}{A numeric vector of lower bounds.}
  \item{ub}{A numeric vector of upper bounds.}
  \item{dim}{An integer representing the number of dimensions, 3.}
}
}
\description{
This function return the default value of lower and upper bounds also the dimension for SVR optimization.
The three dimensions represent as the parameter that need to be optimized in SVR with exact range of bound.
Three dimension and the range represent as: Cost (C): 2^0 to 2^10; Gamma: 2^(-8) to 2^0; Epsilon: 2^(-8) to 2^0.
}
\examples{
bounds <- get_default_bounds()
bounds$lb  # Lower bounds
bounds$ub  # Upper bounds
bounds$dim # Number of parameters

}
