% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWO_opt.R
\name{initGWO}
\alias{initGWO}
\title{Initialize Position on Gray Wolf Optimizer}
\usage{
initGWO(N, dim, ub, lb)
}
\arguments{
\item{N}{An integer indicate population size.}

\item{dim}{An integer show the number of dimension (parameters) of the problem to optimize. It indicate the number of parameters to be optimized.}

\item{ub}{A numeric vector that show upper bounds of the search space. One value per dimension}

\item{lb}{A numeric vector that show lower bounds of the search space. One value per dimension.}
}
\value{
A numeric matrix of shape \code{(N, dim)} representing initialized positions.
}
\description{
This function generates the initial position of gray wolf within the defined upper and lower bound in every dimension.
}
\note{
This function used inside GWO function for initialization process.
}
