% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHHOCBO_opt.R
\name{levyEHHOCBO}
\alias{levyEHHOCBO}
\title{Levy Flight Generator}
\usage{
levyEHHOCBO(dim)
}
\arguments{
\item{dim}{An integer that indicate the dimensionality of search space.}
}
\value{
A numeric vector of length \code{dim} representing the Lévy flight step.
}
\description{
Generates a random step vector based on Lévy flight distribution, used in the exploitation phase of HHO that used as combined in EHHOCBO.
}
\note{
This function used inside EHHOCBO function to generate random Levy Flight vector.
}
\keyword{internal}
