% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHO_opt.R
\name{levyHHO}
\alias{levyHHO}
\title{Levy Flight Generator}
\usage{
levyHHO(dim)
}
\arguments{
\item{dim}{An integer that indicate the dimensionality of search space.}
}
\value{
A numeric vector of length \code{dim} representing the Lévy flight step.
}
\description{
Generates a random step vector based on Lévy flight distribution, used in the exploitation phase of the HHO algorithm.
}
\note{
This function used inside HHO function to generate random Levy Flight vector.
}
