% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(preds, actuals)
}
\arguments{
\item{preds}{A numeric vector of predicted values.}

\item{actuals}{A numeric vector of actual (true) values.}
}
\value{
MAE value.
}
\description{
Calculate the RMSE value between predicted and actual values.
}
\examples{
preds <- c(80, 120, 180)
actuals <- c(95, 115, 177)
mae(preds, actuals)

}
