% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error}
\usage{
mape(preds, actuals)
}
\arguments{
\item{preds}{A numeric vector of predicted values.}

\item{actuals}{A numeric vector of actual (true) values.}
}
\value{
MAPE value (percentage).
}
\description{
Calculate the MAPE value between predicted and actual values.
Can't be used if the actual values contain 0 value.
}
\examples{
preds <- c(80, 120, 180)
actuals <- c(95, 115, 177)
mape(preds, actuals)

}
