% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalization.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{is a predictor variable that is a numeric vector to be normalized.}
}
\value{
A numeric vector scaled between 0 and 1.
}
\description{
Normalize data using min-max scale.
}
\examples{
# Normalize example use:
data <- c(10, 20, 30, 40, 50)
normalize(data)

}
