% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALO_opt.R
\name{Random_walk_around_antlion}
\alias{Random_walk_around_antlion}
\title{Perform Random Walk Around Antlion}
\usage{
Random_walk_around_antlion(dim, Max_iter, lb, ub, antlion, current_iter)
}
\arguments{
\item{dim}{An integer show the number of dimension (parameters) of the problem to optimize. It indicate the number of parameters to be optimized.}

\item{Max_iter}{An integer indicate maximum number of iterations.}

\item{lb}{A numeric vector that show lower bounds of the search space. One value per dimension.}

\item{ub}{A numeric vector that show upper bounds of the search space. One value per dimension.}

\item{antlion}{A numeric vector representing the position of the selected antlion.}

\item{current_iter}{The current iteration count.}
}
\value{
A numeric matrix of shape \code{(N, dim)} representing the position of the ant in each step of the random walk.
}
\description{
Function simulates random walk of an ant within the boundaries influenced by an antlion's position.
}
\note{
This function used inside ALO function to update the position of ants.
}
