% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(preds, actuals)
}
\arguments{
\item{preds}{A numeric vector of predicted values.}

\item{actuals}{A numeric vector of actual (true) values.}
}
\value{
RMSE value.
}
\description{
Calculate the RMSE value between predicted and actual values.
}
\examples{
preds <- c(80, 120, 180)
actuals <- c(95, 115, 177)
rmse(preds, actuals)

}
