% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function.R
\name{smape}
\alias{smape}
\title{Symmetric Mean Absolute Percentage Error}
\usage{
smape(preds, actuals)
}
\arguments{
\item{preds}{A numeric vector of predicted values.}

\item{actuals}{A numeric vector of actual (true) values.}
}
\value{
SMAPE value (percentage).
}
\description{
Calculate the SMAPE value between predicted and actual values.
}
\examples{
preds <- c(80, 120, 180)
actuals <- c(95, 115, 177)
smape(preds, actuals)

}
